requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
numpy==1.9.2
py2neo==2.0.6
python-igraph==0.7.1.post4
scipy==0.15.1
Werkzeug==0.10.4
